libhb: correct pgs aspect ratio during passthrough
The PGS subtitle stream format frame size is encoded in the stream.
Since the frame aspect ratio is usually changed due to bounds clipping
while converting a video, the resulting aspect ratio is incorrect and
players incorrectly display the resulting subtitles.

This change processes the PGS stream to correct the frame size and
picture offsets so they match the geometry of the conversion job.

Signed-off-by: Jeremiah McConnell <miah@miah.com>